ignore ETAG query test as well, reuse skip util
